+++ packages/styled-components/src/utils/flatten.js
@@ -8,6 +8,7 @@ import isStyledComponent from './isStyledComponent';
 import Keyframes from '../models/Keyframes';
 import hyphenate from './hyphenateStyleName';
 import addUnitIfNeeded from './addUnitIfNeeded';
+import { type Stringifier } from '../types';
 
 /**
  * It's falsish not falsy because 0 is allowed.
@@ -15,38 +16,39 @@ import addUnitIfNeeded from './addUnitIfNeeded';
 const isFalsish = chunk => chunk === undefined || chunk === null || chunk === false || chunk === '';
 
 export const objToCssArray = (obj: Object, prevKey?: string): Array<string | Function> => {
-  const rules = []
-  const keys = Object.keys(obj)
+  const rules = [];
+  const keys = Object.keys(obj);
 
-  keys
-    .forEach(key => {
-      if (!isFalsish(obj[key])) {
-        if (isPlainObject(obj[key])) {
-          rules.push(...objToCssArray(obj[key], key))
+  keys.forEach(key => {
+    if (!isFalsish(obj[key])) {
+      if (isPlainObject(obj[key])) {
+        rules.push(...objToCssArray(obj[key], key));
 
-          return rules
-        }
-        else if (isFunction(obj[key])) {
-          rules.push(`${hyphenate(key)}:`, obj[key], ';')
+        return rules;
+      } else if (isFunction(obj[key])) {
+        rules.push(`${hyphenate(key)}:`, obj[key], ';');
 
-          return rules
-        }
-        rules.push(`${hyphenate(key)}: ${addUnitIfNeeded(key, obj[key])};`);
+        return rules;
       }
-      return rules;
-    })
+      rules.push(`${hyphenate(key)}: ${addUnitIfNeeded(key, obj[key])};`);
+    }
+    return rules;
+  });
 
-    return prevKey
-      ? [`${prevKey} {`, ...rules, '}']
-      : rules;
+  return prevKey ? [`${prevKey} {`, ...rules, '}'] : rules;
 };
 
-export default function flatten(chunk: any, executionContext: ?Object, styleSheet: ?Object): any {
+export default function flatten(
+  chunk: any,
+  executionContext: ?Object,
+  styleSheet: ?Object,
+  stylisInstance: ?Stringifier
+): any {
   if (Array.isArray(chunk)) {
     const ruleSet = [];
 
     for (let i = 0, len = chunk.length, result; i < len; i += 1) {
-      result = flatten(chunk[i], executionContext, styleSheet);
+      result = flatten(chunk[i], executionContext, styleSheet, stylisInstance);
 
       if (result === '') continue;
       else if (Array.isArray(result)) ruleSet.push(...result);
@@ -79,14 +81,14 @@ export default function flatten(chunk: any, executionContext: ?Object, styleShee
         );
       }
 
-      return flatten(result, executionContext, styleSheet);
+      return flatten(result, executionContext, styleSheet, stylisInstance);
     } else return chunk;
   }
 
   if (chunk instanceof Keyframes) {
     if (styleSheet) {
-      chunk.inject(styleSheet);
-      return chunk.getName();
+      chunk.inject(styleSheet, stylisInstance);
+      return chunk.getName(stylisInstance);
     } else return chunk;
   }